requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
requests>=2.31.0
python-dotenv>=1.0.0
asyncpg>=0.28.0
python-multipart>=0.0.6
aiofiles>=23.2.0
elevenlabs>=2.16.0
pydantic>=2.0.0